Added frontend and backend for assignment. Update and delete has been implemented. Error handeling still need to be done

+++ api/controllers/assignmentController.js
@@ -9,7 +9,7 @@ const mongoose = require("mongoose");
 const Assignments = require("../models/courses/assigment"),
 	Subjects = require("../models/courses/subject"),
 	Files = require("../models/file");
-const {remove_files} = require("../helpers/re_useables/reuse");
+const { remove_files } = require("../helpers/re_useables/reuse");
 
 //# Get all request
 
@@ -19,7 +19,7 @@ exports.assignments_list = async (req, res) => {
 		fullpath = res.fullpath;
 
 	try {
-		assigment = await Assignments.find().populate("file", "_id url");
+		assigment = await Assignments.find().populate("file", "_id name url");
 
 		// If none exist
 		if (!assigment.length >= 1) return res.status(200).json({ message: "No Assigments exist", assigments: assigment });
@@ -27,12 +27,11 @@ exports.assignments_list = async (req, res) => {
 		let response = {
 			message: "Feched all data",
 			count: assigment.length,
-			id: req.params,
 			assigments: assigment.map(doc => {
 				return {
 					_id: doc._id,
 					title: doc.title,
-					description: doc.type,
+					description: doc.description,
 					type: doc.type,
 					file: doc.file,
 					subject: doc.subject,
@@ -42,102 +41,113 @@ exports.assignments_list = async (req, res) => {
 					changed_at: doc.changed_at,
 					request: {
 						type: "GET",
-						url: origin.slice(-1) === "/" ? `${fullpath}${doc._id}` : `${fullpath}/${doc._id}`
-					}
+						url: origin.slice(-1) === "/" ? `${fullpath}${doc._id}` : `${fullpath}/${doc._id}`,
+					},
 				};
-			})
+			}),
 		};
 
 		res.status(200).json(response);
 	} catch (error) {
 		res.status(500).json({
 			message: "Something went wrong",
-			error: error.message
+			error: error.message,
 		});
 	}
 };
 
 exports.assignments_create = async (req, res, next) => {
-	// use params subject
 	try {
-		let files = req.files,
-			teacher = res.userData,
-			files_id = [],
-			sub_id = req.params.sub_id;	
-		if (files.length >= 1) {
-			for (let file of files) {
-				new_file = await new Files({
-					name: file.filename,
-					url: file.path,
-					ext: file.filename.split(".").pop(),
-					type: file.mimetype
-				}).save();
+		let files = req.files;
+		teacher = res.userData;
+		files_id = [];
 
-				files_id.push(new_file._id);
+		if (files) {
+			if (files.length >= 1) {
+				for (let file of files) {
+					file_path = res.hostname + String(file.path).replace(/\\/g, "/").slice(6);
+					new_file = await new Files({
+						name: file.filename,
+						path: file.path,
+						url: file_path,
+						ext: file.filename.split(".").pop(),
+						type: file.mimetype,
+					}).save();
+
+					files_id.push(new_file._id);
+				}
 			}
 		}
+
 		let file = files_id.length >= 1 ? files_id : [];
+		subject = req.body.subject ? req.body.subject : null;
 
 		let new_assigment = await new Assignments({
 			title: req.body.title,
 			description: req.body.description,
 			type: req.body.type,
 			file: file,
-			subject: sub_id,
-			// subject: req.body.subject,
-			teacher: teacher.id,
-			deadline: req.body.deadline
+			subject,
+			// teacher: teacher.id,
+			deadline: req.body.deadline,
 		}).save();
 
 		res.status(201).json({
 			message: "Assigment Created",
-			assigment: new_assigment
+			assigment: new_assigment,
 		});
 	} catch (error) {
 		res.status(500).json({
-			message: "Something went wrong",
-			error: error.message
+			message: "Something went wrong 1",
+			error: error.message,
 		});
 	}
 };
 
 // todo permission checking
 exports.assigments_update = async (req, res, next) => {
-	let id = req.params.id;
-	let files = req.files;
 	// let teacher = res.userData; // todo teacher's ID
-	let files_id = [];
 
-	//# files
-	// - find assignment first
+	let id = req.params.id,
+		files = req.files,
+		replace_current = req.body.replaceCurrent,
+		files_id = [];
+	// convert to array
 
-	task = await Assignments.findById(id).populate("file", "_id url ");
+	// files
+	//  find assignment first
+	//todo maak een uitzondering op bestanden dat niet gewijzigd zijn
+
+	task = await Assignments.findById(id).populate("file", "_id path ");
 	task_files = task.file;
+	// //! check of bestand aanwezig is in de niet gewijzigd lijst
 
 	// - Delete exsisting files in public folder
-	if (files.length >= 1) {
-		let errors = [];
-		// todo reuse gebruiken
-		try {
-			task_files.forEach(async file => {
-				if (fs.existsSync(file.url)) {
-					// - Delete file in public folder
-					fs.unlink(file.url, error => {
-						if (error) errors.push({ message: error });
-					});
-				} else {
-					errors.push({ message: `File: ${file.url} doesn't exist` });
-				}
-			});
-		} catch (error) {
-			errors.push({ message: error });
-		}
-		if (errors.length >= 1) {
-			//! misschien toegevoegde bestanden ook mee verwijderen?
-			return res.status(500).json({
-				message: "Something went wrong",
-				err: errors
-			});
+	if (replace_current != "undefined" && String(replace_current).toLowerCase() === "true") {
+		if (files.length >= 1) {
+			let errors = [];
+			// todo reuse gebruiken
+			try {
+				task_files.forEach(async file => {
+					if (fs.existsSync(file.path)) {
+						// - Delete file in public folder
+						fs.unlink(file.path, error => {
+							if (error) errors.push({ message: error });
+						});
+					} else {
+						errors.push({ message: `File: ${file.path} doesn't exist` });
+					}
+				});
+			} catch (error) {
+				errors.push({ message: error.message });
+			}
+			if (errors.length >= 1) {
+				//! misschien toegevoegde bestanden ook mee verwijderen?
+				return res.status(500).json({
+					message: "Something went wrong 1",
+					err: errors,
+				});
+			}
 		}
 	}
 
@@ -146,43 +156,57 @@ exports.assigments_update = async (req, res, next) => {
 		// //- Adding new (updated) file if any
 		let old_files_id = task_files.map(file => file._id);
 		if (files.length >= 1) {
-			await Files.deleteMany({ _id: { $in: old_files_id } });
-			file_path =res.hostname + String(file.path).replace(/\\/g, "/").slice(6);
+			if (replace_current != "undefined" && String(replace_current).toLowerCase() === "true") {
+				try {
+					await Files.deleteMany({ _id: { $in: old_files_id } });
+				} catch (error) {
+					return res.status(500).json({
+						message: "Something went wrong 2",
+						error: error.message,
+					});
+				}
+			}
+
 			for (let file of files) {
+				url = res.hostname + String(file.path).replace(/\\/g, "/").slice(6);
 				new_file = await new Files({
 					name: file.filename,
-					url: file_path,
+					path: file.path,
+					url,
 					ext: file.filename.split(".").pop(),
-					type: file.mimetype
+					type: file.mimetype,
 				}).save();
+
 				files_id.push(new_file._id);
 			}
 		}
-
-		//# Assignment
-		// Updating Assignment
-		//! misschien files leeglaten als niks opgegeven is
+		// //# Assignment
+		// // Updating Assignment
+		// //! misschien files leeglaten als niks opgegeven is
 		let file = files_id.length >= 1 ? files_id : old_files_id;
-		var update_task = {
+		if (replace_current != "undefined" && String(replace_current).toLowerCase() === "false") {
+			file.push(...old_files_id);
+		}
+
+		let update_task = {
 			title: req.body.title,
 			description: req.body.description,
 			type: req.body.type,
 			file: file,
-			// subject: req.body.subject,
+			subject: req.body.subject,
 			// teacher: teacher.id,
-			deadline: req.body.deadline
+			deadline: req.body.deadline,
 		};
 		updated_assignment = await Assignments.findByIdAndUpdate(id, update_task, { useFindAndModify: false });
-
 		return res.status(200).json({
 			message: "Updating Assigment",
 			current_assignment: task,
-			updated_assignment: update_task
+			updated_assignment: update_task,
 		});
 	} catch (error) {
 		return res.status(500).json({
-			message: "Something went wrong",
-			error: error
+			message: "Something went wrong 3",
+			error: error.message,
 		});
 	}
 };
@@ -190,23 +214,22 @@ exports.assigments_update = async (req, res, next) => {
 //-error handling
 exports.assigments_delete = async (req, res, next) => {
 	let id = req.params.id;
-	let files_ids = req.body.file;
-	let only_files = req.body.file_only;
+	let files_ids = req.body.files_ids;
+	let delete_task = req.body.delete_task;
 	if (!(files_ids instanceof Array)) {
 		if (typeof files_ids === "undefined") files_ids = [];
 		else files_ids = new Array(files_ids);
 	}
+	if (!(await Assignments.findById(id))) return res.status(404).json({ message: "NO assignment with such ID" });
 
-	if (!await Assignments.findById(id)) return res.status(404).json({ message: "NO assignment with such ID" });
-
-	//- delete  assignment only if only_files (only files should be deleted and not the assigments itself)
-	if (only_files === "false") {
+	// //- delete  assignment only if only_files (only files should be deleted and not the assigments itself)
+	if (delete_task) {
 		try {
 			await Assignments.findByIdAndDelete(id);
 		} catch (error) {
 			return res.status(500).json({
-				message: "Something went wrong",
-				err: error
+				message: "Something went wrong 1",
+				err: error.message,
 			});
 		}
 	}
@@ -216,14 +239,13 @@ exports.assigments_delete = async (req, res, next) => {
 		errors = await remove_files({ files_ids: files_ids, Model_ref: Assignments, _id: id });
 	}
 	if (errors.length >= 1) {
-		//! misschien toegevoegde bestanden ook mee verwijderen?
 		return res.status(500).json({
-			message: "Something went wrong",
-			error: errors
+			message: "Something went wrong 2",
+			error: errors,
 		});
 	}
 	return res.status(202).json({
-		message: "Removal succesfull"
+		message: "Removal succesfull",
 	});
 };
